+++ empulse/metrics/metric/__init__.py
@@ -2,7 +2,19 @@ from .base_metric import BaseMetric
 from .cost_matrix import CostMatrix
 from .metric import Metric
 from .mixture_metric import MixtureComponent, MixtureMetric
-from .strategies import AUEPC, Cost, EmpiricalMaxProfit, LogCost, LogitObjective, MaxProfit, MetricStrategy, Savings
+from .strategies import (
+    AUEPC,
+    Cost,
+    EmpiricalMaxProfit,
+    EmpiricalMinCost,
+    LogCost,
+    LogitObjective,
+    MaxProfit,
+    MetricStrategy,
+    MinCost,
+    Profit,
+    Savings,
+)
 
 __all__ = [
     'AUEPC',
@@ -10,12 +22,15 @@ __all__ = [
     'Cost',
     'CostMatrix',
     'EmpiricalMaxProfit',
+    'EmpiricalMinCost',
     'LogCost',
     'LogitObjective',
     'MaxProfit',
     'Metric',
     'MetricStrategy',
+    'MinCost',
     'MixtureComponent',
     'MixtureMetric',
+    'Profit',
     'Savings',
 ]