+++ empulse/metrics/metric/strategies/__init__.py
@@ -1,9 +1,21 @@
 from .auepc_strategy import AUEPC
-from .cost_strategy import Cost
-from .empirical_max_profit_strategy import EmpiricalMaxProfit
+from .cost_strategy import Cost, Profit
+from .empirical_max_profit_strategy import EmpiricalMaxProfit, EmpiricalMinCost
 from .log_cost_strategy import LogCost
-from .max_profit_strategy import MaxProfit
+from .max_profit_strategy import MaxProfit, MinCost
 from .metric_strategy import LogitObjective, MetricStrategy
 from .savings_strategy import Savings
 
-__all__ = ['AUEPC', 'Cost', 'EmpiricalMaxProfit', 'LogCost', 'LogitObjective', 'MaxProfit', 'MetricStrategy', 'Savings']
+__all__ = [
+    'AUEPC',
+    'Cost',
+    'EmpiricalMaxProfit',
+    'EmpiricalMinCost',
+    'LogCost',
+    'LogitObjective',
+    'MaxProfit',
+    'MetricStrategy',
+    'MinCost',
+    'Profit',
+    'Savings',
+]