+++ empulse/metrics/metric/strategies/max_profit_strategy/__init__.py
@@ -1,3 +1,3 @@
-from .max_profit_strategy import MaxProfit
+from .max_profit_strategy import MaxProfit, MinCost
 
-__all__ = ['MaxProfit']
+__all__ = ['MaxProfit', 'MinCost']